+++ addOns/jython/src/main/java/org/zaproxy/zap/extension/jython/ExtensionJython.java
@@ -54,6 +54,7 @@ public class ExtensionJython extends ExtensionAdaptor {
 
     private ExtensionScript extScript = null;
     private JythonOptionsParam jythonOptionsParam;
+    private JythonOptionsPanel jythonOptionsPanel;
 
     public ExtensionJython() {
         super(NAME);
@@ -71,14 +72,25 @@ public class ExtensionJython extends ExtensionAdaptor {
                         new JythonEngineWrapper(jythonOptionsParam, new PyScriptEngineFactory()));
 
         extensionHook.addOptionsParamSet(this.jythonOptionsParam);
-        if (null != super.getView()) {
-            extensionHook.getHookView().addOptionPanel(new JythonOptionsPanel());
+        if (hasView()) {
+            String[] scriptEngineNode = {
+                Constant.messages.getString("options.script.title"),
+                Constant.messages.getString("scripts.options.engine.title")
+            };
+            getView().getOptionsDialog().addParamPanel(scriptEngineNode, getOptionsPanel(), true);
         }
     }
 
     @Override
     public boolean canUnload() {
-        return false;
+        return true;
+    }
+
+    @Override
+    public void unload() {
+        if (hasView()) {
+            getView().getOptionsDialog().removeParamPanel(getOptionsPanel());
+        }
     }
 
     private ExtensionScript getExtScript() {
@@ -90,6 +102,13 @@ public class ExtensionJython extends ExtensionAdaptor {
                                     .getExtension(ExtensionScript.NAME);
         }
         return extScript;
+    }
+
+    private JythonOptionsPanel getOptionsPanel() {
+        if (jythonOptionsPanel == null) {
+            jythonOptionsPanel = new JythonOptionsPanel();
+        }
+        return jythonOptionsPanel;
     }
 
     @Override