+++ addOns/jython/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 ### Changed
 - Update minimum ZAP version to 2.14.0.
+- Move "Jython" under "Scripts > Engine" in the Options panel list.
 
 ## [13] - 2023-09-07
 ### Changed